feat(Classe Quarto) terminei todo o processo
Classe Quarto redefinida para não possuir os Getters, apenas os atributos, o construtor e as funções - Arquivo CPP de quarto recebendo os dados, verificando se ele existe e armazenando eles em um arquivo quartos.txt

+++ Header-files/Quarto.h
@@ -17,7 +17,7 @@ class Quarto {
     
     //Getters
     public:
-      Quarto(int hospedes, int diaria, bool status);
+      Quarto(int hospedes, float diaria, bool status);
       bool quartoExiste(string identificador);
       virtual string criaStringDeDados();
       virtual int armazenaDadosEmArquivo(string dados);
